requirements.txt
# ============================================================
# OCRExtract — Python dependencies
# Python 3.10+ required
# ============================================================

# ── Web framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Settings / config ─────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Google Drive API ──────────────────────────────────────
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ── PDF → Image ───────────────────────────────────────────
# Requires Poppler binaries on PATH (see README)
pdf2image==1.17.0
Pillow==10.3.0

# ── OCR ───────────────────────────────────────────────────
# Requires Tesseract binary on PATH (see README)
pytesseract==0.3.10

# ── Image preprocessing ───────────────────────────────────
opencv-python-headless==4.9.0.80
numpy==1.26.4

# ── Text / fuzzy matching ─────────────────────────────────
rapidfuzz==3.9.3

# ── Excel output ──────────────────────────────────────────
pandas==2.2.2
openpyxl==3.1.2

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
